+++ tbot/machine/linux/auth.py
@@ -51,3 +51,11 @@ class PrivateKeyAuthenticator(Authenticator):
         :param pathlib.Path key_file: Path to the private key file
         """
         self.key_file = key_file
+
+
+class NoneAuthenticator(Authenticator):
+    """
+    Authenticates without key or password.
+
+    Useful if an ssh-agent is running in the background on lab-hosts.
+    """